moved function ShowHidePlans from App.tsx to its own file and installed packages for new imports
